moved ModelLabelsAPI code into predict.py
