+++ api/__init__.py
@@ -1,3 +1,2 @@
 from .metadata import ModelMetadataAPI  # noqa
-from .labels import ModelLabelsAPI  # noqa
-from .predict import ModelPredictAPI  # noqa
+from .predict import ModelLabelsAPI, ModelPredictAPI  # noqa
